Cleaned up cloud build by renaming files and moving methods around (#219)
* Cleaned up cloud build

* Fixed bug

* More renaming and comment cleanup

* Added helper file

* Fixed missing return
