+++ .cloud-build/execute_notebook_remote.py
@@ -1,3 +1,20 @@
+#!/usr/bin/env python
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Methods to run a notebook on Google Cloud Build"""
+
 from google.protobuf import duration_pb2
 from yaml.loader import FullLoader
 
@@ -22,8 +39,7 @@ def execute_notebook_remote(
     container_uri: str,
     tag: Optional[str],
 ) -> operation.Operation:
-    """Create and execute a simple Google Cloud Build configuration,
-    print the in-progress status and print the completed status."""
+    """Create and execute a single notebook on Google Cloud Build"""
 
     # Authorize the client with Google defaults
     credentials, project_id = google.auth.default()
